+++ src/Rule/Digit.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Stadly\PasswordPolice\Rule;
 
+use Stadly\PasswordPolice\Constraint\Count;
 use Stadly\PasswordPolice\Policy;
 
 final class Digit extends CharacterClass
@@ -11,52 +12,53 @@ final class Digit extends CharacterClass
     /**
      * @param int $min Minimum number of digits.
      * @param int|null $max Maximum number of digits.
+     * @param int $weight Constraint weight.
      */
-    public function __construct(int $min = 1, ?int $max = null)
+    public function __construct(int $min = 1, ?int $max = null, int $weight = 1)
     {
-        parent::__construct('0123456789', $min, $max);
+        parent::__construct('0123456789', $min, $max, $weight);
     }
 
     /**
      * {@inheritDoc}
      */
-    protected function getMessage(): string
+    protected function getMessage(Count $constraint, int $count): string
     {
         $translator = Policy::getTranslator();
 
-        if ($this->max === null) {
+        if ($constraint->getMax() === null) {
             return $translator->trans(
                 'There must be at least one digit.|'.
                 'There must be at least %count% digits.',
-                ['%count%' => $this->min]
+                ['%count%' => $constraint->getMin()]
             );
         }
 
-        if ($this->max === 0) {
+        if ($constraint->getMax() === 0) {
             return $translator->trans(
                 'There must be no digits.'
             );
         }
 
-        if ($this->min === 0) {
+        if ($constraint->getMin() === 0) {
             return $translator->trans(
                 'There must be at most one digit.|'.
                 'There must be at most %count% digits.',
-                ['%count%' => $this->max]
+                ['%count%' => $constraint->getMax()]
             );
         }
 
-        if ($this->min === $this->max) {
+        if ($constraint->getMin() === $constraint->getMax()) {
             return $translator->trans(
                 'There must be exactly one digit.|'.
                 'There must be exactly %count% digits.',
-                ['%count%' => $this->min]
+                ['%count%' => $constraint->getMin()]
             );
         }
 
         return $translator->trans(
             'There must be between %min% and %max% digits.',
-            ['%min%' => $this->min, '%max%' => $this->max]
+            ['%min%' => $constraint->getMin(), '%max%' => $constraint->getMax()]
         );
     }
 }